requirements.txt
# Core Django
Django==4.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Database
mysqlclient==2.2.7
django-mysql==4.11.0

# API
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7

# Real-time & Async
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File Handling
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# Geolocation
django-geojson==3.2.0
geopy==2.4.1

# Search
django-haystack==3.2.1
whoosh==2.7.4

# Email & Notifications
django-allauth==0.57.0
celery==5.3.4
redis==5.0.1

# Security
django-axes==6.0.0
django-defender==1.0.0

# Monitoring
django-debug-toolbar==4.2.0
django-silk==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Development
ipython==8.17.2
black==23.11.0
flake8==6.1.0

# Desktop App Dependencies (for PyQt option)
# PyQt6==6.6.0
# PyQt6-WebEngine==6.6.0

# Mobile (Future)
# react-native requirements would be separate
